Overhaul CMakeLists, create include directory
Change CMakeLists so picotft can both be included
as a subdirectory and built as its own project.
Add a copy of pico_sdk_import.cmake so the
pico-sdk can be found when the library is built by
itself.
Enable many compiler warnings.
Move public headers to a new include directory.