Overhaul DisplayIO with PIO and DMA
Add gen_cmd_write.pio PIO program to send commands
and data to the display without need for processor
bit-banging.
Use DMA to transfer data to the PIO's TX queue so
the processor can be even less involved.
Create write request/ticketing system in DisplayIO
to control the asynchronous process of sending
data to the display.
Probably break a bunch of things because this
commit hasn't been tested.
Add padding if necessary to Renderer pixel buffer
so the DMA can use the more efficient 32-bit
transfer mode.
Add DisplayCmdWrite struct that holds parameters
to the new method DisplayIO::writeCmdEx.
Add writeCmdEx because the number of parameters I
was stuffing into the signature of writeCmd to
control DMA stuff and callbacks was getting
ridiculous.
Add UpdatableDisplayByte to manage Display
properties that are sent to the device when
updated.

Move headers for rendering stuff to
include/picotft/rendering.